fix PP2PM module warning
