+++ pandapower/opf/validate_opf_input.py
@@ -1,6 +1,5 @@
 import pandas as pd
 
-
 def _check_necessary_opf_parameters(net, logger):
     # Check if all necessary parameters are given:
     opf_col = {
